Update README.md
This project is designed to automate the testing of critical functionalities on the Automaticity website, specifically focusing on registration, login, API interactions, and the shopping workflow. The test framework leverages Selenium WebDriver for UI automation and Postman for API validation, ensuring that all key features work as expected.

The test scenarios cover both positive and negative test cases, verifying:

The ability to register and log in with valid and invalid credentials.
Edge cases, such as boundary values and missing inputs.
The correct behavior of password recovery and logout functionalities.
API responses and interactions with different endpoints.
The end-to-end shopping workflow to ensure a smooth user experience.
By automating these scenarios, the project enhances testing efficiency, reduces manual effort, and ensures the stability of the website’s core features. The test scripts are structured following the Page Object Model (POM) for better maintainability and scalability.